Transaction controller authorization, transaction policies, transaction requests cleaned up, casts, added tests

+++ app/Http/Controllers/TransactionController.php
@@ -17,8 +17,13 @@ use Illuminate\Routing\Redirector;
 
 class TransactionController extends Controller
 {
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function index(Wallet $wallet): Factory|View|Application
     {
+        $this->authorize('view', [Transaction::class, $wallet]);
+
         $transactions = $wallet->transactions()->latest()->get();
         $totalIncoming = $wallet->transactions()->where('is_incoming', true)->sum('amount');
         $totalOutgoing = $wallet->transactions()->where('is_incoming', false)->sum('amount');
@@ -31,12 +36,17 @@ class TransactionController extends Controller
         return view('transactions.create', compact('wallet'));
     }
 
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function store(
         TransactionStoreRequest $request,
         Wallet $wallet,
         TransactionService $service
     ): Redirector|Application|RedirectResponse
     {
+        $this->authorize('create', [Transaction::class, $wallet]);
+
         $data = new TransactionCreate(...$request->validated());
         $service->create($wallet, $data);
 
@@ -44,18 +54,28 @@ class TransactionController extends Controller
             ->with('alert', ['type' => 'success', 'message' => 'Transaction created successfully']);
     }
 
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function update(
         TransactionUpdateRequest $request,
         Transaction $transaction
     ): Response|Redirector|RedirectResponse {
+        $this->authorize('update', [$transaction, $transaction->wallet]);
+
         $transaction->fill($request->validated())->save();
 
-        return redirect(route('wallets.transactions.index', $transaction->wallet_id))
+        return redirect(route('wallets.transactions.index', $transaction->wallet))
             ->with('alert', ['type' => 'success', 'message' => 'Transaction updated successfully']);
     }
 
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function destroy(Transaction $transaction, TransactionService $service): Redirector|Application|RedirectResponse
     {
+        $this->authorize('delete', [$transaction, $transaction->wallet]);
+
         $service->delete($transaction);
 
         return redirect(route('wallets.transactions.index', $transaction->wallet))
